% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpp.R
\name{dist}
\alias{dist}
\title{Calculate Euclidean distance between two vectors}
\usage{
dist(v1, v2)
}
\arguments{
\item{v1}{First vector}

\item{v2}{Second vector}
}
\value{
The Euclidean distance
}
\description{
Calculate Euclidean distance between two vectors
}
