% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpp.R
\name{kpp}
\alias{kpp}
\title{Calculate initial centers for K-means clustering using the K-means++ algorithm}
\usage{
kpp(data, k)
}
\arguments{
\item{data}{An n x 2 data.frame}

\item{k}{The desired number of centers to return}
}
\value{
A data.frame of centers
}
\description{
Calculate initial centers for K-means clustering using the K-means++ algorithm
}
\examples{
d = faithful \%>\% rename(x = eruptions, y = waiting)
kpp_centers = kpp(d,2)
kmeans(d,centers = kpp_centers)
}
